requirements.txt
django
djangorestframework
djangorestframework-jwt
django-rest-auth
django-allauth
coreapi